requirements-all.txt
# =====================================
# FORGE AI - Consolidated Requirements
# Frontend Scanner + Backend Generator
# =====================================

# Core LangChain and LangGraph
langchain==1.1.2
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.1
langchain-openai==1.1.0
langchain-groq>=0.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.12
langsmith==0.4.55

# AI/ML Libraries
sentence-transformers>=5.0.0
openai==2.9.0
tiktoken==0.12.0

# Vector Database
chromadb==1.3.5
faiss-cpu==1.13.1

# Web Frameworks
fastapi==0.123.9
starlette==0.50.0
uvicorn==0.38.0
httptools==0.7.1
watchfiles==1.1.1
websockets==15.0.1

# HTTP Client Libraries
httpx==0.28.1
httpcore==1.0.9
httpx-sse==0.4.3
aiohttp==3.13.2
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0

# Database ORMs
sqlalchemy==2.0.44
alembic==1.13.1
prisma==0.12.0

# Data Validation
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5

# Code Generation and Parsing
jinja2==3.1.3
black==24.2.0
autopep8==2.0.4
isort==5.13.2
libcst==1.1.0
# ast-grep==0.18.0  # Not available via pip, install separately if needed
tree-sitter==0.25.2
tree-sitter-javascript==0.25.0
tree-sitter-typescript==0.23.2

# File Handling
aiofiles==25.1.0
pathspec==0.12.1
gitignore_parser==0.1.13

# CLI and UI
click==8.3.1
rich==14.2.0
typer==0.20.0
typer-slim==0.20.0
coloredlogs==15.0.1
colorama==0.4.6
humanfriendly==10.0

# Configuration
python-dotenv==1.2.1
pyyaml==6.0.3

# Data Serialization
orjson==3.11.4
ormsgpack==1.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
marshmallow==3.26.1
dataclasses-json==0.6.7

# Infrastructure
kubernetes==34.1.0
docker==7.0.0
grpcio==1.76.0

# Git Integration
gitpython==3.1.42

# Authentication
bcrypt==5.0.0
oauthlib==3.3.1

# Google Cloud
google-auth==2.43.0
googleapis-common-protos==1.72.0

# Monitoring/Telemetry
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
posthog==5.4.0

# Utilities
backoff==2.2.1
tenacity==9.1.2
durationpy==0.10
distro==1.9.0
annotated-doc==0.0.4
annotated-types==0.7.0
overrides==7.7.0
mypy_extensions==1.1.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0

# Data Processing
numpy==2.3.5
mmh3==5.2.0
xxhash==3.6.0

# Networking
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
anyio==4.12.0
sniffio==1.3.1
websocket-client==1.9.0
h11==0.16.0

# Compression/Encoding
zstandard==0.25.0
pybase64==1.4.2

# File System
fsspec==2025.12.0
hf-xet==1.2.0

# Build Tools
build==1.3.0
pyproject_hooks==1.2.0

# Core Dependencies
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
urllib3==2.3.0
filelock==3.20.0
packaging==25.0
six==1.17.0
attrs==25.4.0
frozenlist==1.8.0
multidict==6.7.0
propcache==0.4.1
yarl==1.22.0
cachetools==6.2.2
greenlet==3.3.0

# Runtime/System
onnxruntime==1.23.2
flatbuffers==25.9.23
sympy==1.14.0
mpmath==1.3.0

# Markdown/Text
markdown-it-py==4.0.0
mdurl==0.1.2
pygments==2.19.2

# Database Tools
pypika==0.48.9

# Date/Time
python-dateutil==2.9.0.post0

# Async/Concurrency
importlib_metadata==8.7.0
importlib_resources==6.5.2
jiter==0.12.0

# Utils
uuid_utils==0.12.0
referencing==0.37.0
rpds-py==0.30.0
zipp==3.23.0
regex==2025.11.3
tqdm==4.67.1
pyreadline3==3.5.4

# Security
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1
shellingham==1.5.4

# HuggingFace
huggingface_hub>=0.20.0
tokenizers==0.22.1

# =====================================
# Development Dependencies (Optional)
# =====================================

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3

# Type Stubs
types-PyYAML==6.0.12.12
types-requests==2.31.0.20240125

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Development Tools
ipython==8.21.0
ipdb==0.13.13
